requirements.txt
llama-index
llama-index-llms-openai
llama-index-llms-gemini
llama-index-embeddings-gemini
google-generativeai
llama-index-vector-stores-weaviate
python-dotenv
matplotlib
weaviate-client
ragas
llama-index-packs-corrective-rag